% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangular.R
\name{clipExcel}
\alias{clipExcel}
\title{Clip Excel: Copy Tibble to Clipboard in Excel Compatible Format}
\usage{
clipExcel(x, row.names = FALSE, col.names = TRUE, na = "", ...)
}
\arguments{
\item{x}{A tibble.}

\item{row.names}{Logical, include row names?}

\item{col.names}{Logical, include column names?}

\item{na}{What to do with NA values? Default to blank}

\item{...}{Any further arguments to write.table}
}
\description{
The clipExcel function takes in a tibble / dataframe and arguments for row
and column names. Its only return value is that tibble copied to the
clipboard in a format that can be easily pasted to Excel
Taken from \url{https://stackoverflow.com/questions/24704344/copy-an-r-data-frame-to-an-excel-spreadsheet}
}
\examples{
\donttest{ clipExcel(iris) }
}
\keyword{tibble}
