% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciences.R
\name{compareSeq}
\alias{compareSeq}
\title{Compare Sequences}
\usage{
compareSeq(ref, seq)
}
\arguments{
\item{ref}{Reference}

\item{seq}{Sequence to compare}
}
\value{
3 column tibble of Position, Ref, and ALT
}
\description{
Produces a table showing all the variants between two DNAString objects
}
