% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangular.R
\name{exposeDupes}
\alias{exposeDupes}
\title{Expose Duplicates}
\usage{
exposeDupes(x, grouping_var, listout = TRUE)
}
\arguments{
\item{x}{Tibble or Dataframe}

\item{grouping_var}{Column to look for duplicated values}

\item{listout}{Flag to return either list or tibble}
}
\value{
List or Dataframe of results
}
\description{
Sometimes datasets are expected to be tidy but aren't, finding distinct
rows of duplicated IDs is easy but finding why they're distinct in many
column tables is less straight forward. This functions returns the values
that resulted in any duplicated IDs in one of two forms either a named list
or a tibble
}
\details{
Named list of two-column tibbles for each value resulting in duplicate IDs
\itemize{
 \item{Grouping Variable}
 \item{Distinct values}
 }

Tibble with the following columns
\itemize{
 \item{Grouping Variable}
 \item{n * X.grpNdistinct} {number of distinct values for duplicated ID}
 \item{n * X.values} {values for that duplicated ID}
}

https://dplyr.tidyverse.org/articles/programming.html
}
\examples{
\donttest{
  library(tidyselect)
  df <- data.frame(name = sample(letters, 20, replace = TRUE),
               month = sample(month.name, 20, replace = TRUE),
            letters = sample(LETTERS[1:10], 20, replace = TRUE),
            nums = floor(runif(20, 1, 15)))
  dplyr::count(df, name)
  exposeDupes(df, name)
  exposeDupes(df, name, listout = FALSE)
  }
}
