% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n, interleave = TRUE)
}
\arguments{
\item{n}{Number of colours in palette}

\item{interleave}{Interleave or not}
}
\description{
Create a ggplot palette and optionally interleave high-low-luminance
}
\examples{
gg_color_hue(n = 10 )
}
