% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not In}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Vector of values}

\item{y}{Another vector of values}
}
\value{
Vector of values that are in X but not in Y
}
\description{
Opposite of \%in\%
}
\examples{
\dontrun{
 x \%!in\% y
 }
}
