% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Vector of values}

\item{y}{Another vector of values}
}
\value{
Vector of logical values inverting X %in% Y
}
\description{
Opposite of %in%
}
\examples{
\donttest{
 names(iris) \%notin\% "Species"
 }
}
