% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insertClipExcelAddin}
\alias{insertClipExcelAddin}
\title{Insert \%>\% clipExcel() at cursor position}
\usage{
insertClipExcelAddin()
}
\description{
Call this function as an addin to insert \code{ \%>\% clipExcel()} at the cursor position.
Useful for quickly checking pipeline status. Bind to ctrl+shift+V
}
