% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insertGlimpseAddin}
\alias{insertGlimpseAddin}
\title{Insert \%>\% glimpse() at cursor position}
\usage{
insertGlimpseAddin()
}
\description{
Call this function as an addin to insert \code{ \%>\% glimpse()} at the cursor position.
Useful for quickly checking pipeline status. Bind to ctrl+shift+V
}
