% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insertViewAddin}
\alias{insertViewAddin}
\title{Insert \%>\% View() at cursor position}
\usage{
insertViewAddin()
}
\description{
Call this function as an addin to insert \code{ \%>\% View()} at the cursor position.
Useful for quickly checking pipeline status. Bind to ctrl+shift+V
}
