% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{interleave}
\alias{interleave}
\title{interleave}
\usage{
interleave(v1, v2)
}
\arguments{
\item{v1}{Vector 1}

\item{v2}{Vector 2}
}
\description{
Interleave two vectors
}
\examples{
interleave(v1 = c(1,2,3) , v2 = c(4,5,6) )
}
