% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x, newlims = c(0, 1))
}
\arguments{
\item{x}{Vector of values}

\item{newlims}{numeric vector of new minimum and maximum eg. c(2,4)}
}
\value{
Normalized vector of values to newmax and newmin
}
\description{
Normalizes a vector of values to a range of 0-1
x - min(x)) / (max(x) - min(x)
}
\examples{
normalize(mtcars$mpg)
normalize(mtcars$mpg, c(5, 10))
}
