% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ops}
\alias{ops}
\title{Plain English CIGAR Operation Names}
\format{
A tibble with 9 rows and 2 variables:
\describe{
  \item{operation}{CIGAR codes: M I D N S H P = X}
  \item{Operation}{Translated codes: Match Insertion Deletion Skipped Soft Clip Hard Clip Padding Sequence Match Sequence Mismatch}
  ...
}
}
\source{
\url{https://drive5.com/usearch/manual/cigar.html}
}
\usage{
ops
}
\description{
A dataset containing the translations from codes to english for CIGAR operations
}
\keyword{datasets}
