% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciences.R
\name{plot_all_cigars}
\alias{plot_all_cigars}
\title{Plot All CIGARs}
\usage{
plot_all_cigars(x, lg = F)
}
\arguments{
\item{x}{tidied cigar data from tidy_cigar}

\item{lg}{Create large lists: will warn if over 20 reads in SAM. Better to filter prior}
}
\value{
List of Plots
}
\description{
Plot all Queries from tidied cigar table, facet by Reference, and colour by Operation
}
