% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciences.R
\name{read_sam}
\alias{read_sam}
\title{Read SAM}
\usage{
read_sam(file)
}
\arguments{
\item{file}{path to file}
}
\value{
11 column tibble of QNAME, FLAG, RNAME, POS, MAPQ, CIGAR, RNEXT, PNEXT, TLEN, SEQ, QUAL
}
\description{
Read in SAM file as tibble
}
