% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciences.R
\name{read_sam_headers}
\alias{read_sam_headers}
\title{Read SAM Headers}
\usage{
read_sam_headers(file)
}
\arguments{
\item{file}{path to file}
}
\value{
List of two tibbles: References (RNAME/LENGTH) and Mapping (ID/PN/VN/CL)
}
\description{
Read in SAM file headers as list of two tibbles: one with reference names and lengths, the other with mapping info
}
