% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciences.R
\name{sam_to_fasta}
\alias{sam_to_fasta}
\title{Write SAM to Fasta file}
\usage{
sam_to_fasta(x, file)
}
\arguments{
\item{x}{SAM file in tabular format (eg. as read by read_sam())}

\item{file}{Path to output file}
}
\value{
Nothing. File written to disk
}
\description{
Write Query sequences to file as fasta with with QNAME as identifier
}
