% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samcols}
\alias{samcols}
\title{Plain English CIGAR Operation Names}
\format{
An 11-element character vector:
\describe{
  \item{chr}{column names: QNAME FLAG RNAME POS MAPQ CIGAR RNEXT PNEXT TLEN SEQ QUAL}
  ...
}
}
\source{
\url{http://samtools.github.io/hts-specs/SAMv1.pdf}
}
\usage{
samcols
}
\description{
A dataset containing the columns of a SAM file
}
\keyword{datasets}
