% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciences.R
\name{tidy_cigar}
\alias{tidy_cigar}
\title{Tidy CIGAR Data}
\usage{
tidy_cigar(x)
}
\arguments{
\item{x}{SAM in tibble format}
}
\value{
6 column tibble of QNAME, RNAME, MAPQ, CIGARstart, CIGARend, Operation
}
\description{
Split and tidy CIGARS from read_sam's tibble. Groups by QNAME and RNAME.
}
