% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{utilitarian}
\alias{utilitarian}
\title{utilitarian: A package for collecting useful R functions}
\description{
Eventually useful R scripts have repeated portions that deserve
more than just copying and pasting.
}
\section{Package Related}{

Related somehow to other packages
\itemize{
  \item \code{\link{usePackage}}
  \item \code{\link{libraries}}
}
}

