% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colours in palette}
}
\description{
Create an interleaved high-low-luminance pal
}
\examples{
gg_color_hue(n = 10 )
}
