% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-related.R
\name{libraries}
\alias{libraries}
\title{Call multiple library statements at once}
\usage{
libraries(..., verbose = F, quoted = F)
}
\arguments{
\item{...}{Package names}

\item{verbose}{Logical whether to announce via message() what packages are being attached.}

\item{quoted}{Logical whether package names supplied bare or quoted as character}
}
\description{
Attach a number of library packages at once
}
\examples{
libraries(rmarkdown, cranlogs, visdat, rentrez)
libraries("rmarkdown", "cranlogs", "visdat", "rentrez", quoted = T)
libraries(c("broom", "forcats"), quoted = T)
libraries(visdat, verbose = TRUE)
}
\author{
Adrian Zetner
}
\keyword{library}
\keyword{packages}
\keyword{utilitarian}
